Refactor: Split ParticipationsTable component
Split the ParticipationsTable component into smaller, more manageable files to improve code organization and readability.
